chore: add predeploy_address and move it into it's own module
